Support OAuth endpoints which don't return expires_in field
Fixes #297

Previously when using the `OAuthAuthenticator` with an OAuth endpoint
which didn't include the `expires_in` field in the response you'd see
a `KeyError`, e.g.

    File "/.../site-packages/singer_sdk/authenticators.py", line 432, in update_access_token
        self.expires_in = token_json["expires_in"]
    KeyError: 'expires_in'

The `expires_in` field is RECOMMENDED but not REQUIRED[1] - so this is a
use-case we want to handle. To solve the issue we add a
default_expiration argument to `OAuthAuthenticator`, which will be used
as a fallback if the OAuth endpoint does not include an `expires_in`
with its response. This change also comes with a slightly more subtle
side-effect, now if no `default_expiration` is set, and the OAuth
response has no `expires_in` then the authenticator will assume the
token never expires (see the second condition in `is_token_valid()`).

[1]: https://datatracker.ietf.org/doc/html/rfc6749#section-4.2.2