Enable JSONPath for GraphQL record extraction
The current expectation for a graphql response is that its structure
would match the stream name. There are occasions, however, when this is
not the case. In such situations there is currently a need to override
the parse_response method, but it would be more straightforward to use a
consistent approach to the parent RESTStream and simply set the
records_jsonpath variable.

Here we allow the GraphQL stream to inherit parse_response method from the
RESTStream and set the default value to ensure the behaviour matches how
it was before.