fix(targets): Added a condition to the `No schema for record field` warning (#2348)
fix: :bug: add a condition to the `No schema for record field` warning

If schema flattening is enabled but a value for a key in the record which is flattened is None, this warning is returned because the record can not be flattened and then the key is not in the schema. This can be avoided by checking if the value is None. If it is None the record field actually does not exist so we do not need a warning.